Attempting to get a version that will compile.
I've been hunting down bugs just trying to get a base version of the code that compiles. My thinking is that if we get a base version that compiles, and all of a sudden it stops compiling, we don't have to look through so much code the find the errors.

Note that the esos_f14ui.c and t3_app.c both compile to create .o files, but for some reason scons isn't generating the .hex file.

Also note that this code contains Wes' branch's most current code (from Sunday), although it has some issues.